+++ src/index.d.ts
@@ -30,7 +30,6 @@ export interface SignUpData {
   referrer: string;
   brandingId?: number;
   metadata?: { [key: string]: string };
-  dateOfBirth?: string;
 }
 
 export interface RequestPasswordData {
@@ -167,6 +166,7 @@ export interface DirectDebitChargeData {
   assetId: string;
   accessFeeId: string;
   voucherCode: string;
+  brandingId?: number;
 }
 
 export interface DirectDebitChargeResponse {